+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release Notes for Teamleader
 
+- Added fields on Deals
+- Deal Title
+- Deal Value
+
+- Company
+- Made email optional to match API
+- Created option to optionally require VAT number
+
+- Added option to set a size - so when users have a lot of custom fields, they can actually increase the base limit of 20.
+
+- Fixed issue with the API calls on 'context' filters, now we actually only get the fields of correct context through the API.
+
 ## 5.0.2 - 10-03-2025
 ### Fixed
 - Fixed the path of the icon-mask to `teamleader`, using an alias looks to the namespace, not the folder structure or plugin handle